Add RuntimeContext test placeholder, other minor test refactor
