+++ test/OpenTelemetry.Tests/OpenTelemetry.Tests.csproj
@@ -4,11 +4,6 @@
     <TargetFrameworks>netcoreapp2.1;netcoreapp3.1</TargetFrameworks>
     <TargetFrameworks Condition="$(OS) == 'Windows_NT'">$(TargetFrameworks);net46;net452</TargetFrameworks>
   </PropertyGroup>
-  <ItemGroup>
-    <Compile Remove="Implementation\**" />
-    <EmbeddedResource Remove="Implementation\**" />
-    <None Remove="Implementation\**" />
-  </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="$(RepoRoot)\src\OpenTelemetry\OpenTelemetry.csproj" />
@@ -23,9 +18,5 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
     </PackageReference>
     <DotNetCliToolReference Include="dotnet-xunit" Version="$(DotNetXUnitCliVer)" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Testing\Export\" />
   </ItemGroup>
 </Project>
